+++ docs/CHANGELOG.md
@@ -1,5 +1,12 @@
 # GoNB Changelog
 
+
+## Next
+
+* Added a `replace` directive from `go.lsp.dev/{jsonrpc2,uri,protocol}` to `github.com/go-language-server/{jsonrpc2,uri,protocol}`
+  in `go.mod`, since `go.lsp.dev` seems to be down.
+* Updated dependencies to `go.lsp.dev/uri` and `go.lsp.dev/protocol`.
+
 ## v0.10.7, 2025/01/20
 
 * Improved `autostart` logic: it now requires being mounted as "readonly" under `/root/autostart`.